Split off Metadata
