+++ src/plugins/text-mode/aug/rawToAug.ts
@@ -1,8 +1,8 @@
 import Aug from "./AugState";
 import * as Graph from "@desmodder/graph-state";
-import Metadata from "main/metadata/interface";
-import migrateToLatest from "main/metadata/migrate";
 import { ChildExprNode, evalMaybeRational, AnyNode } from "parsing/parsenode";
+import Metadata from "plugins/manage-metadata/interface";
+import migrateToLatest from "plugins/manage-metadata/migrate";
 import { parseDesmosLatex } from "utils/depUtils";
 
 export default function rawToAug(raw: Graph.GraphState): Aug.State {
